brookyapi/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.3
dataclasses_json
numpy
google-cloud-storage
google-api-core
currencyconverter
pandas
gymnasium==0.28.1
dask
dask[dataframe]
stable_baselines3
sb3-contrib
torch